repo: move `StoreFactories::default()` etc. to separate module
This is high-level glue code that we want to remove from the `repo`
module, so that module can be moved into a lower-level `jj-core`
crate.
