+++ lib/tests/test_init.rs
@@ -189,7 +189,7 @@ fn test_init_load_non_utf8_path() -> TestResult {
     use std::ffi::OsStr;
     use std::os::unix::ffi::OsStrExt as _;
 
-    use jj_lib::workspace::default_working_copy_factories;
+    use jj_lib::default_backend_factories::default_working_copy_factories;
     use pollster::FutureExt as _;
     use testutils::TestEnvironment;
 
@@ -219,7 +219,7 @@ fn test_init_load_non_utf8_path() -> TestResult {
     let workspace = Workspace::load(
         &settings,
         &workspace_root,
-        &test_env.default_store_factories(),
+        &test_env.default_backend_factories(),
         &default_working_copy_factories(),
     )?;
