+++ lib/tests/test_local_working_copy_concurrent.rs
@@ -16,10 +16,10 @@ use std::cmp::max;
 use std::thread;
 
 use assert_matches::assert_matches;
+use jj_lib::default_backend_factories::default_working_copy_factories;
 use jj_lib::repo::Repo as _;
 use jj_lib::working_copy::CheckoutError;
 use jj_lib::workspace::Workspace;
-use jj_lib::workspace::default_working_copy_factories;
 use pollster::FutureExt as _;
 use testutils::TestResult;
 use testutils::TestWorkspace;
@@ -59,7 +59,7 @@ fn test_concurrent_checkout() -> TestResult {
         let mut ws2 = Workspace::load(
             &settings,
             &workspace1_root,
-            &test_workspace1.env.default_store_factories(),
+            &test_workspace1.env.default_backend_factories(),
             &default_working_copy_factories(),
         )?;
         // Reload commit from the store associated with the workspace
@@ -80,7 +80,7 @@ fn test_concurrent_checkout() -> TestResult {
     let ws3 = Workspace::load(
         &settings,
         &workspace1_root,
-        &test_workspace1.env.default_store_factories(),
+        &test_workspace1.env.default_backend_factories(),
         &default_working_copy_factories(),
     )?;
     assert_tree_eq!(*ws3.working_copy().tree()?, tree2);
@@ -125,7 +125,7 @@ fn test_checkout_parallel() -> TestResult {
                 let mut workspace = Workspace::load(
                     &settings,
                     &workspace_root,
-                    &test_env.default_store_factories(),
+                    &test_env.default_backend_factories(),
                     &default_working_copy_factories(),
                 )
                 .unwrap();
